backend/structure-service/pyproject.toml
[tool.poetry]
name = "structure-service"
version = "0.1.0"
description = ""
authors = ["irumako <irumakoTS@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
python-keycloak = "^4.0.0"
pydantic-settings = "^2.2.1"
neontology = "^1.3.1"
neo4j = "^5.20.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
